MIC-2524: Update mortality approach per model (#15)
This change updates the simulation to made a change in the model detailed in:

https://github.com/ihmeuw/vivarium_research/pull/520/files

Specifically, it replaces the CSMR for multiple myeloma with ACMR and replaces the EMR (0) for the susceptible state with ACMR - CSMR_c486.

Also, this change includes a burn-in period of 10 years (2011-2021) and corrects misspellings of isatuximab and datatumumab.

Tested with a remade artifact, full parallel run and results processing. Output looked as expected using the survival analysis notebook and perusing the csvs in count space. Note: other_cause deaths are counted as deaths by the susceptible state due to a workaround for mortality data (susceptible having an EMR).