Resolve issue resulting in 0 accounting in susceptible state (#40)
There was a bug in the results stratifier where records whose values that are only set at diagnosis (i.e., `undiagnosed`) were not counted in results. This meant that everyone who was in the susceptible state (by definition undiagnosed and not with-condition) were left out of stratification. This issue did not affect survival or registry results as those data already are limited to those outside of the susceptible state. 

Tested with a single simulate run and results processing.